+++ R/zzz.R
@@ -6,6 +6,7 @@
   register_s3_method("dplyr", "db_query_fields", "BigQueryConnection")
   register_s3_method("dplyr", "db_save_query", "BigQueryConnection")
   register_s3_method("dplyr", "sql_translate_env", "BigQueryConnection")
+  register_s3_method("dbplyr", "db_copy_to", "BigQueryConnection")
 
   # Default options --------------------------------------------------------
   op <- options()